navbar update with different components
