--- app/bus-rental/page.tsx
@@ -1,20 +1,0 @@
-'use client';
-
-import React from 'react';
-
-const BusRentalPage: React.FC = () => {
-  return (
-    <div
-      style={{
-        fontFamily: 'Arial, sans-serif',
-        textAlign: 'center',
-        marginTop: '100px',
-        color: '#000'
-      }}
-    >
-      <h1>This is dedicated for BUS RENTAL</h1>
-    </div>
-  );
-};
-
-export default BusRentalPage;
