+++ app/dashboard/bus-assignment/page.tsx
@@ -1,0 +1,16 @@
+'use client';
+
+import React from 'react';
+
+const BusAssignmentPage: React.FC = () => {
+  return (
+    <div className="dashboard-content">
+      <div className="center-box">
+        This is dedicated for BUS ASSIGNMENT
+      </div>
+    </div>
+  );
+};
+
+export default BusAssignmentPage;
+
