+++ app/dashboard/bus-operation/page.tsx
@@ -1,0 +1,15 @@
+'use client';
+
+import React from 'react';
+
+const BusOperationPage: React.FC = () => {
+  return (
+    <div className="dashboard-content">
+      <div className="center-box">
+        This is dedicated for BUS OPERATION
+      </div>
+    </div>
+  );
+};
+
+export default BusOperationPage;
