+++ app/dashboard/bus-rental/page.tsx
@@ -1,0 +1,15 @@
+'use client';
+
+import React from 'react';
+
+const BusRentalPage: React.FC = () => {
+  return (
+    <div className="dashboard-content">
+      <div className="center-box">
+        This is dedicated for BUS RENTAL
+      </div>
+    </div>
+  );
+};
+
+export default BusRentalPage;
