+++ app/dashboard/gps/page.tsx
@@ -1,0 +1,15 @@
+'use client';
+
+import React from 'react';
+
+const GPSPage: React.FC = () => {
+  return (
+    <div className="dashboard-content">
+      <div className="center-box">
+        This is dedicated for GPS
+      </div>
+    </div>
+  );
+};
+
+export default GPSPage;
