+++ app/dashboard/performance-report/page.tsx
@@ -1,0 +1,15 @@
+'use client';
+
+import React from 'react';
+
+const PerformanceReportPage: React.FC = () => {
+  return (
+    <div className="dashboard-content">
+      <div className="center-box">
+        This is dedicated for PERFORMANCE REPORT
+      </div>
+    </div>
+  );
+};
+
+export default PerformanceReportPage;
