+++ app/dashboard/qouta-assignment/page.tsx
@@ -1,0 +1,15 @@
+'use client';
+
+import React from 'react';
+
+const QuotaAssignmentPage: React.FC = () => {
+  return (
+    <div className="dashboard-content">
+      <div className="center-box">
+        This is dedicated for QOUTA ASSIGNMENT
+      </div>
+    </div>
+  );
+};
+
+export default QuotaAssignmentPage;
