+++ app/dashboard/route-management/page.tsx
@@ -1,0 +1,15 @@
+'use client';
+
+import React from 'react';
+
+const RouteManagementPage: React.FC = () => {
+  return (
+    <div className="dashboard-content">
+      <div className="center-box">
+        This is dedicated for Route Management
+      </div>
+    </div>
+  );
+};
+
+export default RouteManagementPage;
