--- app/gps/page.tsx
@@ -1,21 +1,0 @@
-'use client';
-
-import React from 'react';
-
-const GPSPage: React.FC = () => {
-  return (
-    <div
-      style={{
-        fontFamily: 'Arial, sans-serif',
-        textAlign: 'center',
-        marginTop: '100px',
-        color: '#000'
-      }}
-    >
-      <h1>This is dedicated for GPS</h1>
-    </div>
-  );
-};
-
-export default GPSPage;
-
