--- app/performance-report/page.tsx
@@ -1,20 +1,0 @@
-'use client';
-
-import React from 'react';
-
-const PerformanceReportPage: React.FC = () => {
-  return (
-    <div
-      style={{
-        fontFamily: 'Arial, sans-serif',
-        textAlign: 'center',
-        marginTop: '100px',
-        color: '#000'
-      }}
-    >
-      <h1>This is dedicated for PERFORMANCE REPORT</h1>
-    </div>
-  );
-};
-
-export default PerformanceReportPage;
